+++ include/countly.hpp
@@ -56,6 +56,8 @@ public:
 
   void setSha256(cly::SHA256Function fun);
 
+  void enableManualSessionControl();
+
   void setHTTPClient(HTTPClientFunction fun);
 
   void setMetrics(const std::string &os, const std::string &os_version, const std::string &device, const std::string &resolution, const std::string &carrier, const std::string &app_version);
@@ -104,6 +106,11 @@ public:
    * Checks and returns the size of the event queue in memory or persistent storage.
    */
   int checkEQSize();
+
+  /*
+   * Checks and returns the size of the request queue in memory or persistent storage.
+   */
+  int checkRQSize();
 
   /**
    * Checks and returns the size of the event queue in persistent storage.
@@ -319,6 +326,7 @@ private:
   std::chrono::system_clock::duration getSessionDuration();
 
   void updateLoop();
+  void packEvents();
   bool began_session = false;
   bool is_being_disposed = false;
   bool is_sdk_initialized = false;